--- src/components/mainPageLayout/CategoryBar.tsx
@@ -1,1 +1,0 @@
-export {}
